Window level higher than normal
When displaying the Toast under status bar the windowLevel is set to `UIWindowLevelNormal`. However, this is not fully correct because this window is on top of the main application window. When using this library with others like `SVProgressHUD`, they have problems identifying the proper place to insert themselves. Just adding a small amount (+1) works and configures this window on proper level (above the app).